pyproject.toml
[tool.poetry]
name = "django-starter"
version = "0.1.0"
description = "Djang Starter for Alphabell Projects"
authors = ["Asanka Lihiniyagoda"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"
whitenoise = "^6.2.0"
dj-database-url = "^1.0.0"
rollbar = "^0.16.3"
django-storages = "^1.13.1"
redis = "^4.3.4"
celery = {extras = ["redis"], version = "^5.2.7"}
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pre-commit = "^2.20.0"
notebook = "^6.4.12"
django-extensions = "^3.2.0"
flake8 = "^5.0.4"
flake8-isort = "^4.2.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.5.0"
pytest-django = "^4.5.2"
mypy = "^0.971"
django-stubs = "^1.12.0"
pytest = "^7.1.2"
pytest-sugar = "^0.9.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
py_version=38